fix python 3.6 native CentOS 7 support in 123.09beta01
- CentOS 7.7 now as native python 3.6 yum packages named in format python3-, while previously Centmin Mod used python36u from IUS Community and then replaced that with python36 paackages from EPEL. As such yum updates may report conflicts.
- This update attempts to fix those conflicts, when you run cmupdate to update 123.09beta01 and then run centmin.sh menu option once and then exit via centmin.sh menu option 24 - which will prompt you to run /usr/local/src/centminmod/addons/python36_install.sh if python36u is detected. Running /usr/local/src/centminmod/addons/python36_install.sh will then resolve most of known conflicts by replacing python36u or python36 packages with native python3 packages which seem to be available for both CentOS 7.6 and 7.7 now.
- You can also do this by running cmupdate and then /usr/local/src/centminmod/addons/python36_install.sh for 90% of the fixes. For 100% of fixes, cmupdate + centmin.sh menu run once is needed, before running /usr/local/src/centminmod/addons/python36_install.sh
- These conflicts only occur on existing Centmin Mod 123.09beta01 installs and not fresh new ones which will install by default python3 correct native CentOS 7 packages now instead of python36u from IUS Community or python36 from EPEL.

+++ addons/python36_install.sh
@@ -102,24 +102,16 @@ starttime=$(TZ=UTC date +%s.%N)
 {
 
 check_pythonthree_six() {
-  if [[ "$CENTOS_SEVEN" -eq '7' && "$(yum list python36 -q 2>&1 | grep -o No)" = 'No' ]]; then
-    # workaround for premature EPEL python36 package retirement before CentOS 7.7 is ready
-    # https://community.centminmod.com/threads/18142/
-    if [ ! -f "${DIR_TMP}/python36-libs-3.6.8-1.el7.x86_64.rpm" ]; then
-      wget -4 -q -O "${DIR_TMP}/python36-libs-3.6.8-1.el7.x86_64.rpm" https://centminmod.com/centminmodparts/epel/el7/x86_64/python36-libs-3.6.8-1.el7.x86_64.rpm
+  if [[ "$CENTOS_SEVEN" -eq '7' ]]; then
+  # if [[ "$(echo "$CENTOSVER" | sed -e 's|\.||g')" -ge '77' ]]; then
+    # CentOS 7.7+ already have native python 3.6 yum packages
+    # via python3 and python3-libs so no longer require EPEL python36 packages
+    if [[ "$CENTOS_SEVEN" -eq '7' && -z "$(rpm -qa python3)" ]]; then
+      yum -q -y install python3
     fi
-    if [ ! -f "${DIR_TMP}/python36-3.6.8-1.el7.x86_64.rpm" ]; then
-      wget -4 -q -O "${DIR_TMP}/python36-3.6.8-1.el7.x86_64.rpm" https://centminmod.com/centminmodparts/epel/el7/x86_64/python36-3.6.8-1.el7.x86_64.rpm
+    if [[ "$CENTOS_SEVEN" -eq '7' && -z "$(rpm -qa python3-libs)" ]]; then
+      yum -q -y install python3-libs
     fi
-    yum -y localinstall "${DIR_TMP}/python36-3.6.8-1.el7.x86_64.rpm" "${DIR_TMP}/python36-libs-3.6.8-1.el7.x86_64.rpm"
-  fi
-  if [[ "$CENTOS_SEVEN" -eq '7' && "$(yum list python36-libs -q 2>&1 | grep -o No)" = 'No' ]]; then
-    # workaround for premature EPEL python36 package retirement before CentOS 7.7 is ready
-    # https://community.centminmod.com/threads/18142/
-    if [ ! -f "${DIR_TMP}/python36-libs-3.6.8-1.el7.x86_64.rpm" ]; then
-      wget -4 -q -O "${DIR_TMP}/python36-libs-3.6.8-1.el7.x86_64.rpm" https://centminmod.com/centminmodparts/epel/el7/x86_64/python36-libs-3.6.8-1.el7.x86_64.rpm
-    fi
-    yum -y localinstall "${DIR_TMP}/python36-libs-3.6.8-1.el7.x86_64.rpm"
   fi
 }
 
@@ -174,24 +166,64 @@ cecho "*************************************************" $boldgreen
 cecho "Installing Python 3.6" $boldgreen
 cecho "*************************************************" $boldgreen
 
-# install Python 3.6 besides system default Python 2.6
-yum -y install python36u python36u-devel python36u-pip python36u-setuptools python36u-tools --enablerepo=ius
-rpm -ql python36u python36u-devel python36u-pip python36u-setuptools python36u-tools python36u-tkinter | grep bin
-
-# switch in favour of epel python36 version
-# only apply to centos 7 as centos 6 epel doesn't have python36
-if [[ -f /bin/systemctl && "$(rpm -qa python36u)" ]]; then
-  # remove ius community python36u
-  yum -y remove python36u python36u-devel python36u-pip python36u-setuptools python36u-tools python36u-libs python36u-tkinter
-  # install epel python36
-  yum -y install python36 python36-devel python36-pip python36-setuptools python36-tools python36-libs python36-tkinter
+if [[ "$CENTOS_SEVEN" -eq '7' ]]; then
+# if [[ "$(echo "$CENTOSVER" | sed -e 's|\.||g')" -ge '77' ]]; then
+  # prefer CentOS 7.7+ native python3 packages for python 3.6
+    if [[ -f /bin/systemctl && "$(rpm -qa python36u)" ]]; then
+      # remove ius community python36u
+      yum -y remove python36u python36u-devel python36u-pip python36u-setuptools python36u-tools python36u-libs python36u-tkinter
+      if [[ ! "$(rpm -qa cmake3)" || ! "$(rpm -qa cmake3-data)" ]]; then
+        check_pythonthree_six
+        # reinstall removed dependencies from above removed ius community packages
+        yum -y install cmake3 cmake3-data
+      fi
+    fi
+    if [[ -f /bin/systemctl && "$(rpm -qa python36)" ]]; then
+      # remove epel python36
+      yum -y remove python36 python36-devel python36-pip python36-setuptools python36-tools python36-libs python36-tkinter
+      if [[ ! "$(rpm -qa cmake3)" || ! "$(rpm -qa cmake3-data)" ]]; then
+        check_pythonthree_six
+        # reinstall removed dependencies from above removed ius community packages
+        yum -y install cmake3 cmake3-data
+      fi
+    fi
+    yum -y install python3 python3-devel python3-pip python3-setuptools python3-tools python3-libs python3-tkinter
+    rpm -ql python3 python3-devel python3-pip python3-setuptools python3-tools python3-tkinter | grep bin
+    if [[ ! "$(rpm -qa cmake3)" || ! "$(rpm -qa cmake3-data)" ]]; then
+      # reinstall removed dependencies from above removed ius community packages
+      yum -y install cmake3 cmake3-data
+    fi
+# elif [[ "$(echo "$CENTOSVER" | sed -e 's|\.||g')" -lt '77' ]]; then
+#   # install Python 3.6 besides system default Python 2.6
+#   if [[ -f /bin/systemctl && -z "$(rpm -qa python36u)" ]] && [[ -z "$(rpm -qa python36)" ]]; then
+#     # only install python36u if python36 isn't installed
+#     yum -y install python36u python36u-devel python36u-pip python36u-setuptools python36u-tools --enablerepo=ius
+#     rpm -ql python36u python36u-devel python36u-pip python36u-setuptools python36u-tools python36u-tkinter | grep bin
+#   elif [[ -f /bin/systemctl && "$(rpm -qa python36)" && -z "$(rpm -qa python36-tools)" ]]; then
+#     # install epel python36
+#     yum -y install python36 python36-devel python36-pip python36-setuptools python36-tools python36-libs python36-tkinter
+#     rpm -ql python36 python36-devel python36-pip python36-setuptools python36-tools python36-tkinter | grep bin
+#   elif [[ -f /bin/systemctl && -z "$(rpm -qa python36)" ]]; then
+#     # install epel python36
+#     yum -y install python36 python36-devel python36-pip python36-setuptools python36-tools python36-libs python36-tkinter
+#     rpm -ql python36 python36-devel python36-pip python36-setuptools python36-tools python36-tkinter | grep bin
+#   fi
+  
+#   # switch in favour of epel python36 version
+#   # only apply to centos 7 as centos 6 epel doesn't have python36
+#   if [[ -f /bin/systemctl && "$(rpm -qa python36u)" ]]; then
+#     # remove ius community python36u
+#     yum -y remove python36u python36u-devel python36u-pip python36u-setuptools python36u-tools python36u-libs python36u-tkinter
+#     # install epel python36
+#     yum -y install python36 python36-devel python36-pip python36-setuptools python36-tools python36-libs python36-tkinter
+#     rpm -ql python36 python36-devel python36-pip python36-setuptools python36-tools python36-tkinter | grep bin
+#   fi
+#   if [[ ! "$(rpm -qa cmake3)" || ! "$(rpm -qa cmake3-data)" ]]; then
+#     check_pythonthree_six
+#     # reinstall removed dependencies from above removed ius community packages
+#     yum -y install cmake3 cmake3-data
+#   fi
 fi
-if [[ ! "$(rpm -qa cmake3)" || ! "$(rpm -qa cmake3-data)" ]]; then
-  check_pythonthree_six
-  # reinstall removed dependencies from above removed ius community packages
-  yum -y install cmake3 cmake3-data
-fi
-
 
 } 2>&1 | tee ${CENTMINLOGDIR}/python36-install_${DT}.log
 
@@ -199,4 +231,4 @@ endtime=$(TZ=UTC date +%s.%N)
 
 INSTALLTIME=$(echo "scale=2;$endtime - $starttime"|bc )
 echo "" >> ${CENTMINLOGDIR}/python36-install_${DT}.log
-echo "Python 3.5 Install Time: $INSTALLTIME seconds" >> ${CENTMINLOGDIR}/python36-install_${DT}.log
+echo "Python 3.6 Install Time: $INSTALLTIME seconds" >> ${CENTMINLOGDIR}/python36-install_${DT}.log
